Switch from using CRC32 to using MD5.
git-svn-id: https://aomushi.ath.cx:43443/svn/morrigan/trunk@1374 2128fabd-03f8-4a31-9ba6-05665b9bae8a

+++ net.sparktank.morrigan.model.media/src/net/sparktank/morrigan/model/media/IMediaItem.java
@@ -1,5 +1,6 @@
 package net.sparktank.morrigan.model.media;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import net.sparktank.morrigan.model.db.IDbItem;
@@ -14,8 +15,8 @@ public interface IMediaItem extends IDbItem {
 	public Date getDateAdded();
 	public boolean setDateAdded(Date dateAdded);
 		
-	public long getHashcode();
-	public boolean setHashcode(long hashcode);
+	public BigInteger getHashcode();
+	public boolean setHashcode(BigInteger hashcode);
 	
 	public Date getDateLastModified();
 	public boolean setDateLastModified(Date lastModified);
